+++ example/lib/main.dart
@@ -1,4 +1,4 @@
-import 'package:bangla_pdf_fixer_example/invoice_generate.dart';
+import 'package:bangla_pdf_fixer_example/preview.dart';
 import 'package:flutter/material.dart';
 
 void main() => runApp(const MyApp());
@@ -14,7 +14,7 @@ class MyApp extends StatelessWidget {
         colorScheme: ColorScheme.fromSeed(seedColor: Colors.blueAccent),
         useMaterial3: true,
       ),
-      home: const PdfInvoicePage(),
+      home: const PdfPreviewScreen(),
     );
   }
 }